move logging events to separate file
